Facade for fleet manager
