+++ Assets/_FlashCrusade/Scripts/Leader.cs
@@ -5,61 +5,31 @@ using UnityEngine;
 
 public class Leader : MonoBehaviour
 {
-	public float fleetPosTick = 0.5f;
-	private float fleetPosTickTimer;
-
-
+	public FleetManagerFacade fleetManager;
 	public Ship ship;
-	public Fleet fleet;
-	public float fleetShipSpacing = 5f;
+
+	[SerializeField] private float fleetShipSpacing = 5f;
 
 	private bool freeFly = false;
 	public bool FreeFlyOn { get { return freeFly; } }
 
 	private void Start()
 	{
-		fleet = new Fleet(this.ship, fleetShipSpacing);
-
-		Ally[] allyComps = FindObjectsByType<Ally>(FindObjectsSortMode.None);
-		fleet.ships = allyComps.Cast<AIAgent>().ToList();
-
-        fleet.UpdateLocalFleetPositions();
-
-        //UpdateFleetMoveTargets();
+		fleetManager = new FleetManagerFacade(this, fleetShipSpacing);
     }
 
 	private void Update()
 	{
-		if(fleetPosTickTimer > 0)
-		{
-			fleetPosTickTimer -= Time.deltaTime;
-		}
-		else
-		{
-
-			fleetPosTickTimer = fleetPosTick;
-		}
-		fleet.UpdateLocalFleetPositions(); // move back into tick later
-	}
-
-	public void UpdateFleetMoveTargets()
-	{
-		for (int i = 0; i < fleet.ships.Count; i++)
-		{
-			Vector2 worldPos = (Vector2)transform.position + fleet.localFleetPositions[i];
-
-			fleet.ships[i].MoveTarget = worldPos;
-		}
+		fleetManager.Update();
 	}
 
 	#region COMMANDS
 
 	public void ToggleFreeFly()
 	{
-		Debug.Log("Toggle free fly");
 		freeFly = !freeFly;
 
-		foreach(AIAgent ship in fleet.ships)
+		foreach(AIAgent ship in fleetManager.fleet.ships)
 		{
 			if(freeFly)
 			{
@@ -78,7 +48,7 @@ public class Leader : MonoBehaviour
 	{
 		if (!Application.isPlaying) return;
 		Gizmos.color = Color.cyan;
-		foreach (Vector2 position in fleet.localFleetPositions)
+		foreach (Vector2 position in fleetManager.fleet.localFleetPositions)
 		{
 			Gizmos.DrawSphere(transform.position + (Vector3)position, 10);
 		}
